add note in case of custom password policy
If the use case for generating a password involves a custom password policy that requires both upper and lower case, this generator isn't guaranteed to satisfy that requirement.

Since there's no simple way to ensure that, beyond adding extra complexity to the API, this note should be enough to inform users of the possibility they may need to perform additional checks on any generated passwords.